% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate the weights for each marginal likelihood}
\usage{
calc_weights(marg_liks)
}
\arguments{
\item{marg_liks}{(non-log) marginal likelihood estimates}
}
\value{
the weight of each marginal likelihood estimate
}
\description{
Calculate the weights for each marginal likelihood
}
\examples{
  marg_liks <- c(0.0001, 0.0002, 0.0003, 0.0004)
  created <- calc_weights(marg_liks)
  expected <- c(0.1, 0.2, 0.3, 0.4)
  testthat::expect_equal(created, expected)
}
\author{
Richel J.C. Bilderbeek
}
