% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_run_mcbette.R
\name{can_run_mcbette}
\alias{can_run_mcbette}
\title{Can 'mcbette' run?}
\usage{
can_run_mcbette()
}
\description{
Can 'mcbette' run?
Will return \link{TRUE} if:
\itemize{
  \item (1) Running on Linux or MacOS
  \item (2) BEAST2 is installed
  \item (3) The BEAST2 NS package is installed
}
}
\author{
Richèl J.C. Bilderbeek
}
