% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beast2_ns_pkg.R
\name{check_beast2_ns_pkg}
\alias{check_beast2_ns_pkg}
\title{Checks if the BEAST2 'NS' package is installed.}
\usage{
check_beast2_ns_pkg()
}
\description{
Checks if the BEAST2 'NS' package is installed.
Will \link{stop} if not
}
