% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inference_models.R
\name{check_inference_models}
\alias{check_inference_models}
\title{Check if the \code{inference_model} is a valid BEAUti inference model.}
\usage{
check_inference_models(inference_models)
}
\arguments{
\item{inference_models}{a list of one or more inference models,
as can be created by \link[beautier]{create_inference_model}}
}
\value{
nothing
}
\description{
Calls \code{stop} if not.
}
\examples{
 testthat::expect_silent(
   check_inference_models(
       list(create_inference_model()
     )
   )
 )

 # Must stop on nonsense
 testthat::expect_error(check_inference_models("nonsense"))
 testthat::expect_error(check_inference_models(NULL))
 testthat::expect_error(check_inference_models(NA))
}
\seealso{
Use \link{create_inference_model} to create a valid
  BEAST2 options object
}
\author{
Richel J.C. Bilderbeek
}
