% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mcbette_beast2_options.R
\name{create_mcbette_beast2_options}
\alias{create_mcbette_beast2_options}
\title{Create a \code{beast2_options} structure, with the filenames
indicating \code{mcbette} usage}
\usage{
create_mcbette_beast2_options(rng_seed = NA)
}
\description{
Create a \code{beast2_options} structure, with the filenames
indicating \code{mcbette} usage
}
\seealso{
use \link[beastier]{create_beast2_options}
}
