% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ns_inference_model.R
\name{create_ns_inference_model}
\alias{create_ns_inference_model}
\title{Create an inference model to measure the evidence of.}
\usage{
create_ns_inference_model(
  site_model = beautier::create_jc69_site_model(),
  clock_model = beautier::create_strict_clock_model(),
  tree_prior = beautier::create_yule_tree_prior(),
  mcmc = beautier::create_ns_mcmc()
)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link[beautier]{create_site_model}}}

\item{clock_model}{a clock model,
as returned by \code{\link[beautier]{create_clock_model}}}

\item{tree_prior}{a tree priors,
as returned by \code{\link[beautier]{create_tree_prior}}}

\item{mcmc}{one MCMC.
Use \code{\link[beautier]{create_mcmc}} to create an MCMC.
Use \code{\link[beautier]{create_ns_mcmc}} to create an MCMC
  for a Nested Sampling run.
Use \code{\link[beautier]{check_mcmc}} to check if an MCMC is valid.
Use \code{\link[beautier]{rename_mcmc_filenames}} to rename the filenames in an MCMC.}
}
\description{
Create an inference model to measure the evidence of.
To do so, the inference model is created as usual (see
\link[beautier]{create_inference_model}), except
for using a Nested Sampling MCMC (see \link[beautier]{create_ns_mcmc})
}
\examples{
if (can_run_mcbette()) {

  library(testthat)

  inference_model <- create_ns_inference_model()
  expect_true(beautier::is_nested_sampling_mcmc(inference_model$mcmc))
}
}
\author{
Richèl J.C. Bilderbeek
}
