% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_ns_inference_model.R
\name{create_test_ns_inference_model}
\alias{create_test_ns_inference_model}
\title{Create an inference model to be tested by Nested Sampling}
\usage{

  create_test_ns_inference_model(site_model = beautier::create_jc69_site_model(),
  clock_model = beautier::create_strict_clock_model(),
  tree_prior = beautier::create_yule_tree_prior(),
  mcmc = beautier::create_test_ns_mcmc())
}
\arguments{
\item{site_model}{a site model,
as can be created by \link[beautier]{create_site_model}}

\item{clock_model}{a clock model,
as can be created by \link[beautier]{create_clock_model}}

\item{tree_prior}{a tree prior,
as can be created by \link[beautier]{create_tree_prior}}

\item{mcmc}{an MCMC for the Nested Sampling run,
as can be created by \link[beautier]{create_mcmc_nested_sampling}}
}
\description{
Create an inference model to be tested by Nested Sampling
}
