% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_ns_mcmc.R
\name{create_test_ns_mcmc}
\alias{create_test_ns_mcmc}
\title{Create a Nested Sampling MCMC to be used in testing}
\usage{
create_test_ns_mcmc()
}
\description{
Create a Nested Sampling MCMC to be used in testing
}
