% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation of general function arguments.
This function does nothing.
It is intended to inherit function argument documentation.}
\usage{
default_params_doc(
  beast2_bin_path,
  beast2_working_dir,
  beast2_options,
  beast2_optionses,
  clock_model,
  clock_models,
  epsilon,
  fasta_filename,
  inference_model,
  inference_models,
  mcmc,
  os,
  rng_seed,
  site_model,
  site_models,
  tree_prior,
  tree_priors,
  verbose
)
}
\arguments{
\item{beast2_bin_path}{path to the the BEAST2 binary file}

\item{beast2_working_dir}{folder in which BEAST2 will run and
produce intermediate files.
By default, this is a temporary folder}

\item{beast2_options}{a \code{beast2_options} structure,
as can be created by \link[beastier]{create_beast2_options}.}

\item{beast2_optionses}{list of one or more \code{beast2_options}
structures,
as can be created by \link[beastier]{create_beast2_options}.
Use of reduplicated plural to achieve difference with
\code{beast2_options}}

\item{clock_model}{a clock model,
as can be created by \link[beautier]{create_clock_model}}

\item{clock_models}{a list of one or more clock models,
as can be created by \link[beautier]{create_clock_models}}

\item{epsilon}{measure of relative accuracy.
Smaller values result in longer, more precise estimations}

\item{fasta_filename}{name of the FASTA file}

\item{inference_model}{an inference model,
as can be created by \link[beautier]{create_inference_model}}

\item{inference_models}{a list of one or more inference models,
as can be created by \link[beautier]{create_inference_model}}

\item{mcmc}{an MCMC for the Nested Sampling run,
as can be created by \link[beautier]{create_mcmc_nested_sampling}}

\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}

\item{rng_seed}{a random number generator seed used for the BEAST2
inference}

\item{site_model}{a site model,
as can be created by \link[beautier]{create_site_model}}

\item{site_models}{a list of one or more site models,
as can be created by \link[beautier]{create_site_models}}

\item{tree_prior}{a tree prior,
as can be created by \link[beautier]{create_tree_prior}}

\item{tree_priors}{a list of one or more tree priors,
as can be created by \link[beautier]{create_tree_priors}}

\item{verbose}{if TRUE show debug output}
}
\description{
Documentation of general function arguments.
This function does nothing.
It is intended to inherit function argument documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
