% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_marg_lik.R
\name{est_marg_lik}
\alias{est_marg_lik}
\title{Estimate the marginal likelihood for a combination of
site, clock and tree models}
\usage{
est_marg_lik(
  fasta_filename,
  inference_model = beautier::create_inference_model(mcmc = beautier::create_ns_mcmc()),
  beast2_options = create_mcbette_beast2_options(),
  os = rappdirs::app_dir()$os
)
}
\arguments{
\item{fasta_filename}{name of the FASTA file}

\item{inference_model}{an inference model,
as can be created by \link[beautier]{create_inference_model}}

\item{beast2_options}{a \code{beast2_options} structure,
as can be created by \link[beastier]{create_beast2_options}.}

\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
}
\value{
a \link{list} showing the estimated marginal likelihoods
(and its estimated error), its items are::
\itemize{
  \item \code{marg_log_lik}: estimated marginal (natural) log likelihood
  \item \code{marg_log_lik_sd}: estimated error of \code{marg_log_lik}
  \item \code{esses} the Effective Sample Size
  \item \code{estimates} the estimated marginal (natural) log likelihoods
  \item \code{trees} the jointly-estimated posterior trees
}
}
\description{
Estimate the marginal likelihood for a combination of
site, clock and tree models
}
\examples{
if (mauricer::is_beast2_ns_pkg_installed()) {

  library(testthat)

  # An example FASTA file
  fasta_filename <- system.file("extdata", "simple.fas", package = "mcbette")

  # A testing inference model with inaccurate (thus fast) marginal
  # likelihood estimation
  inference_model <- create_test_ns_inference_model()

  # Setup the options for BEAST2 to be able to call BEAST2 packages
  beast2_options <- create_mcbette_beast2_options()

  # Estimate the marginal likelihood
  marg_lik <- est_marg_lik(
    fasta_filename = fasta_filename,
    inference_model = inference_model,
    beast2_options = beast2_options
  )

  # Results are:
  # a list ...
  expect_true(is.list(marg_lik))

  # with these elements ...
  expect_true("marg_log_lik" \%in\% names(marg_lik))
  expect_true("marg_log_lik_sd" \%in\% names(marg_lik))
  expect_true("ess" \%in\% names(marg_lik))

  # with these data types ...
  expect_true(is_one_double(marg_lik$marg_log_lik))
  expect_true(is_one_double(marg_lik$marg_log_lik_sd))
  expect_true(is_one_double(marg_lik$ess))

  # with these values
  expect_true(marg_lik$marg_log_lik < 0.0)
  expect_true(marg_lik$marg_log_lik_sd > 0.0)
  expect_true(marg_lik$ess > 0.0)
}
}
\author{
Richèl J.C. Bilderbeek
}
