% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_marg_liks_from_models.R
\name{est_marg_liks_from_models}
\alias{est_marg_liks_from_models}
\title{Estimate the marginal likelihoods for all combinations of
site, clock and tree models}
\usage{
est_marg_liks_from_models(fasta_filename,
  inference_models = list(beautier::create_inference_model()),
  beast2_optionses = list(beastier::create_beast2_options()),
  epsilon = 1e-12, verbose = FALSE)
}
\arguments{
\item{fasta_filename}{name of the FASTA file}

\item{inference_models}{a list of one or more inference models,
as can be created by \link[beautier]{create_inference_model}}

\item{beast2_optionses}{list of one or more \code{beast2_options}
structures,
as can be created by \link[beastier]{create_beast2_options}.
Use of reduplicated plural to achieve difference with
\code{beast2_options}}

\item{epsilon}{measure of relative accuracy.
Smaller values result in longer, more precise estimations}

\item{verbose}{if TRUE show debug output}
}
\value{
a \link{data.frame} showing the estimated marginal likelihoods
(and its estimated error) per combination of models. Columns are:
\itemize{
  \item \code{site_model_name}: name of the site model
  \item \code{clock_model_name}: name of the clock model
  \item \code{tree_prior_name}: name of the tree prior
  \item \code{marg_log_lik}: estimated marginal (natural) log likelihood
  \item \code{marg_log_lik_sd}: estimated error of \code{marg_log_lik}
  \item \code{weight}: relative model weight, a value from 1.0 (all
    evidence is in favor of this model combination) to 0.0 (no
    evidence in favor of this model combination)
}
}
\description{
Estimate the marginal likelihoods for all combinations of
site, clock and tree models
}
\author{
Richel J.C. Bilderbeek
}
