% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crown_age.R
\name{get_crown_age}
\alias{get_crown_age}
\title{Obtain the crown age of a phylony}
\usage{
get_crown_age(phylogeny)
}
\arguments{
\item{phylo}{a phylogeny of class \link[ape]{phylo}}
}
\value{
the age of the phylogeny
}
\description{
Obtain the crown age of a phylony
}
\examples{
  phylogeny <- ape::read.tree(text = "((A:1, B:1):1, (C:1, D:1):1);")
  crown_age <- get_crown_age(phylogeny)
  testit::assert(crown_age == 2.0)
}
\author{
Richel J.C. Bilderbeek
}
