% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcbette_state.R
\name{get_mcbette_state}
\alias{get_mcbette_state}
\title{Get the current state of \link{mcbette}}
\usage{
get_mcbette_state()
}
\value{
a \link{list} with the following elements:
\itemize{
  \item{
    beast2_installed
      \link{TRUE} if BEAST2 is installed,
      \link{FALSE} otherwise
  }
  \item{
     ns_installed
      \link{NA} if BEAST2 is not installed.
      \link{TRUE} if the BEAST2 NS package is installed
      \link{FALSE} if the BEAST2 NS package is not installed
  }
}
}
\description{
Get the current state of \link{mcbette}
}
\examples{
library(testthat)

state <- get_mcbette_state()
expect_true("beast2_installed" \%in\% names(state))
expect_true("ns_installed" \%in\% names(state))
}
