% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_bayes_factor.R
\name{interpret_bayes_factor}
\alias{interpret_bayes_factor}
\title{Interpret a Bayes factor}
\usage{
interpret_bayes_factor(bayes_factor)
}
\arguments{
\item{bayes_factor}{Bayes factor to be interpreted}
}
\value{
a string with the interpretation in English
}
\description{
Interpret a Bayes factor, using the interpretation from [1].
}
\details{
\itemize{
  \item [1] H. Jeffreys (1961). The Theory of Probability (3rd ed.).
    Oxford. p. 432
}
}
\examples{
library(testthat)
expect_equal(interpret_bayes_factor(0.5), "in favor of other model")
expect_equal(interpret_bayes_factor(1.5), "barely worth mentioning")
expect_equal(interpret_bayes_factor(8.5), "substantial")
expect_equal(interpret_bayes_factor(12.5), "strong")
expect_equal(interpret_bayes_factor(85.0), "very strong")
expect_equal(interpret_bayes_factor(123.0), "decisive")
}
\author{
Richèl J.C. Bilderbeek
}
