% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_marg_lik_estimates.R
\name{interpret_marg_lik_estimates}
\alias{interpret_marg_lik_estimates}
\title{Interpret the marginal likelihood estimates}
\usage{
interpret_marg_lik_estimates(marg_liks)
}
\arguments{
\item{marg_liks}{the estimated marginal likelihoods (aka evidences),
as created by \link{est_marg_liks}}
}
\description{
Interpret the marginal likelihood estimates
as created by \link{est_marg_liks}.
}
\author{
Richèl J.C. Bilderbeek
}
