% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbette_self_test.R
\name{mcbette_self_test}
\alias{mcbette_self_test}
\title{Performs a minimal \link{mcbette} run}
\usage{
mcbette_self_test(beast2_folder = beastier::get_default_beast2_folder())
}
\arguments{
\item{beast2_folder}{the folder where the BEAST2 is installed.
Note that this is not the folder where the BEAST2 executable is installed:
the BEAST2 executable is in a subfolder.
Use \link[beastier]{get_default_beast2_folder}
  to get the default BEAST2 folder.
Use \link[beastier]{get_default_beast2_bin_path}
  to get the full path to the default BEAST2 executable.
Use \link[beastier]{get_default_beast2_jar_path}
  to get the full path to the default BEAST2 jar file.}
}
\value{
Nothing. Will \link{stop} is 'mcbette' cannot run on a minimal,
standard input.
}
\description{
Performs a minimal \link{mcbette} run
}
