% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate the weights for each marginal likelihood}
\usage{
calc_weights(marg_liks)
}
\arguments{
\item{marg_liks}{(non-log) marginal likelihood estimates}
}
\value{
the weight of each marginal likelihood estimate
}
\description{
Calculate the weights for each marginal likelihood
}
\author{
Richel J.C. Bilderbeek
}
