% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate the weights for each marginal likelihood}
\usage{
calc_weights(marg_liks)
}
\arguments{
\item{marg_liks}{(non-log) marginal likelihood estimates}
}
\value{
the weight of each marginal likelihood estimate,
which will sum up to 1.0
}
\description{
Calculate the weights for each marginal likelihood
}
\examples{
# Evidences (aka marginal likelihoods) can be very small
evidences <- c(0.0001, 0.0002, 0.0003, 0.0004)

# Sum will be 1.0
calc_weights(evidences)
}
\author{
Richèl J.C. Bilderbeek
}
