% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_run_mcbette.R
\name{can_run_mcbette}
\alias{can_run_mcbette}
\title{Can 'mcbette' run?}
\usage{
can_run_mcbette(beast2_folder = beastier::get_default_beast2_folder())
}
\arguments{
\item{beast2_folder}{the folder where the BEAST2 is installed.
Note that this is not the folder where the BEAST2 executable is installed:
the BEAST2 executable is in a subfolder.
Use \link[beastier]{get_default_beast2_folder}
  to get the default BEAST2 folder.
Use \link[beastier]{get_default_beast2_bin_path}
  to get the full path to the default BEAST2 executable.
Use \link[beastier]{get_default_beast2_jar_path}
  to get the full path to the default BEAST2 jar file.}
}
\description{
Can 'mcbette' run?
Will return \link{TRUE} if:
\itemize{
  \item (1) Running on Linux or MacOS
  \item (2) BEAST2 is installed
  \item (3) The BEAST2 NS package is installed
}
}
\examples{
can_run_mcbette()

beastier::check_empty_beaustier_folders()
}
\author{
Richèl J.C. Bilderbeek
}
