% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_marg_liks.R
\name{est_marg_liks}
\alias{est_marg_liks}
\title{Estimate the marginal likelihoods for all combinations of
site, clock and tree models}
\usage{
est_marg_liks(
  fasta_filename,
  inference_models = list(beautier::create_inference_model(mcmc =
    beautier::create_ns_mcmc())),
  beast2_optionses = rep(list(beastier::create_mcbette_beast2_options()), times =
    length(inference_models)),
  verbose = FALSE,
  os = rappdirs::app_dir()$os
)
}
\arguments{
\item{fasta_filename}{name of the FASTA file}

\item{inference_models}{a list of one or more inference models,
as can be created by \link[beautier]{create_inference_model}}

\item{beast2_optionses}{list of one or more \code{beast2_options}
structures,
as can be created by \link[beastier]{create_mcbette_beast2_options}.
Use of reduplicated plural to achieve difference with
\code{beast2_options}}

\item{verbose}{if TRUE show debug output}

\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
}
\value{
a \link{data.frame} showing the estimated marginal likelihoods
(and its estimated error) per combination of models. Columns are:
\itemize{
  \item \code{site_model_name}: name of the site model
  \item \code{clock_model_name}: name of the clock model
  \item \code{tree_prior_name}: name of the tree prior
  \item \code{marg_log_lik}: estimated marginal (natural) log likelihood
  \item \code{marg_log_lik_sd}: estimated error of \code{marg_log_lik}
  \item \code{weight}: relative model weight, a value from 1.0 (all
    evidence is in favor of this model combination) to 0.0 (no
    evidence in favor of this model combination)
}
\itemize{
  \item \link{can_run_mcbette}: see if 'mcbette' can run
  \item \link{est_marg_liks}: estimate multiple marginal likelihood of a
    single inference mode
}
}
\description{
Estimate the marginal likelihoods for all combinations of
site, clock and tree models
}
\examples{
if (can_run_mcbette()) {

  # Use an example FASTA file
  fasta_filename <- system.file("extdata", "simple.fas", package = "mcbette")

  # Create two inference models
  inference_model_1 <- beautier::create_test_ns_inference_model(
    site_model = beautier::create_jc69_site_model()
  )
  inference_model_2 <- beautier::create_test_ns_inference_model(
    site_model = beautier::create_hky_site_model()
  )
  inference_models <- list(inference_model_1, inference_model_2)

  beast2_options <- beastier::create_mcbette_beast2_options()

  # Need as much beast2_optionses as inference models
  beast2_optionses <- list(beast2_options, beast2_options)

  est_marg_liks(
    fasta_filename,
    inference_models = inference_models,
    beast2_optionses = beast2_optionses
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
