% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_marg_liks.R
\name{est_marg_liks}
\alias{est_marg_liks}
\title{Estimate the marginal likelihoods for all combinations of
site, clock and tree models}
\usage{
est_marg_liks(fasta_filename, epsilon = 1e-12, verbose = FALSE)
}
\arguments{
\item{fasta_filename}{name of the FASTA file}

\item{epsilon}{measure of relative accuracy.
Smaller values result in longer, more precise estimations}

\item{verbose}{if TRUE show debug output}
}
\value{
a \link{data.frame} showing the estimated marginal likelihoods
(and its estimated error) per combination of models
}
\description{
Estimate the marginal likelihoods for all combinations of
site, clock and tree models
}
\author{
Richel J.C. Bilderbeek
}
