% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_marg_liks.R
\name{est_marg_liks}
\alias{est_marg_liks}
\title{Estimate the marginal likelihoods for all combinations of
site, clock and tree models}
\usage{
est_marg_liks(fasta_filename,
  site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors(), epsilon = 1e-12,
  verbose = FALSE)
}
\arguments{
\item{fasta_filename}{name of the FASTA file}

\item{site_models}{one or more site models,
as, for example, can be created by \link[beautier]{create_site_models}}

\item{clock_models}{one or more clock models,
as, for example, can be created by \link[beautier]{create_clock_models}}

\item{tree_priors}{one or more tree priors,
as, for example, can be created by \link[beautier]{create_tree_priors}}

\item{epsilon}{measure of relative accuracy.
Smaller values result in longer, more precise estimations}

\item{verbose}{if TRUE show debug output}
}
\value{
a \link{data.frame} showing the estimated marginal likelihoods
(and its estimated error) per combination of models. Columns are:
\itemize{
  \item \code{site_model_name}: name of the site model
  \item \code{clock_model_name}: name of the clock model
  \item \code{tree_prior_name}: name of the tree prior
  \item \code{marg_log_lik}: estimated marginal (natural) log likelihood
  \item \code{marg_log_lik_sd}: estimated error of \code{marg_log_lik}
  \item \code{weight}: relative model weight, a value from 1.0 (all
    evidence is in favor of this model combination) to 0.0 (no
    evidence in favor of this model combination)
}
}
\description{
Estimate the marginal likelihoods for all combinations of
site, clock and tree models
}
\author{
Richel J.C. Bilderbeek
}
