% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_marg_liks.R
\name{est_marg_liks}
\alias{est_marg_liks}
\title{Estimate the marginal likelihoods for all combinations of
site, clock and tree models}
\usage{
est_marg_liks(fasta_filename,
  site_models = beautier::create_site_models(),
  clock_models = beautier::create_clock_models(),
  tree_priors = beautier::create_tree_priors(), epsilon = 1e-12,
  rng_seed = 1, verbose = FALSE, os = rappdirs::app_dir()$os)
}
\arguments{
\item{fasta_filename}{name of the FASTA file}

\item{site_models}{a list of one or more site models,
as can be created by \link[beautier]{create_site_models}}

\item{clock_models}{a list of one or more clock models,
as can be created by \link[beautier]{create_clock_models}}

\item{tree_priors}{a list of one or more tree priors,
as can be created by \link[beautier]{create_tree_priors}}

\item{epsilon}{measure of relative accuracy.
Smaller values result in longer, more precise estimations}

\item{rng_seed}{a random number generator seed used for the BEAST2
inference}

\item{verbose}{if TRUE show debug output}

\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
}
\value{
a \link{data.frame} showing the estimated marginal likelihoods
(and its estimated error) per combination of models. Columns are:
\itemize{
  \item \code{site_model_name}: name of the site model
  \item \code{clock_model_name}: name of the clock model
  \item \code{tree_prior_name}: name of the tree prior
  \item \code{marg_log_lik}: estimated marginal (natural) log likelihood
  \item \code{marg_log_lik_sd}: estimated error of \code{marg_log_lik}
  \item \code{weight}: relative model weight, a value from 1.0 (all
    evidence is in favor of this model combination) to 0.0 (no
    evidence in favor of this model combination)
}
}
\description{
Estimate the marginal likelihoods for all combinations of
site, clock and tree models
}
\examples{
  fasta_filename <- system.file("extdata", "simple.fas", package = "mcbette")

  # Only compare two site models
  df <- est_marg_liks(
    fasta_filename,
    site_models = list(
      beautier::create_jc69_site_model(),
      beautier::create_hky_site_model()
    ),
    clock_models = list(beautier::create_strict_clock_model()),
    tree_priors = list(beautier::create_yule_tree_prior()),
    epsilon = 1e7
  )

  testthat::expect_true(is.data.frame(df))
  testthat::expect_true("site_model_name" \%in\% colnames(df))
  testthat::expect_true("clock_model_name" \%in\% colnames(df))
  testthat::expect_true("tree_prior_name" \%in\% colnames(df))
  testthat::expect_true("marg_log_lik" \%in\% colnames(df))
  testthat::expect_true("marg_log_lik_sd" \%in\% colnames(df))
  testthat::expect_true("weight" \%in\% colnames(df))

  testthat::expect_true(is.factor(df$site_model_name))
  testthat::expect_true(is.factor(df$clock_model_name))
  testthat::expect_true(is.factor(df$tree_prior_name))
  testthat::expect_true(!is.factor(df$marg_log_lik))
  testthat::expect_true(!is.factor(df$marg_log_lik_sd))
  testthat::expect_true(!is.factor(df$weight))

  testthat::expect_true(sum(df$marg_log_lik < 0.0, na.rm = TRUE) > 0)
  testthat::expect_true(sum(df$marg_log_lik_sd > 0.0, na.rm = TRUE) > 0)
  testthat::expect_true(all(df$weight >= 0.0))
  testthat::expect_true(all(df$weight <= 1.0))
}
\author{
Richel J.C. Bilderbeek
}
