% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_marg_liks_from_models.R
\name{est_marg_liks_from_models}
\alias{est_marg_liks_from_models}
\title{Estimate the marginal likelihoods for all combinations of
site, clock and tree models}
\usage{
est_marg_liks_from_models(fasta_filename,
  inference_models = list(beautier::create_inference_model(mcmc =
  beautier::create_nested_sampling_mcmc())),
  beast2_optionses = rep(list(beastier::create_beast2_options()), times =
  length(inference_models)), epsilon = 1e-12, verbose = FALSE)
}
\arguments{
\item{fasta_filename}{name of the FASTA file}

\item{inference_models}{a list of one or more inference models,
as can be created by \link[beautier]{create_inference_model}}

\item{beast2_optionses}{list of one or more \code{beast2_options}
structures,
as can be created by \link[beastier]{create_beast2_options}.
Use of reduplicated plural to achieve difference with
\code{beast2_options}}

\item{epsilon}{measure of relative accuracy.
Smaller values result in longer, more precise estimations}

\item{verbose}{if TRUE show debug output}
}
\value{
a \link{data.frame} showing the estimated marginal likelihoods
(and its estimated error) per combination of models. Columns are:
\itemize{
  \item \code{site_model_name}: name of the site model
  \item \code{clock_model_name}: name of the clock model
  \item \code{tree_prior_name}: name of the tree prior
  \item \code{marg_log_lik}: estimated marginal (natural) log likelihood
  \item \code{marg_log_lik_sd}: estimated error of \code{marg_log_lik}
  \item \code{weight}: relative model weight, a value from 1.0 (all
    evidence is in favor of this model combination) to 0.0 (no
    evidence in favor of this model combination)
}
}
\description{
Estimate the marginal likelihoods for all combinations of
site, clock and tree models
}
\examples{
  # Use an example FASTA file
  fasta_filename <- system.file("extdata", "simple.fas", package = "mcbette")

  # Create two inference models
  inference_model_1 <- beautier::create_inference_model(
    site_model = beautier::create_jc69_site_model(),
    mcmc = beautier::create_nested_sampling_mcmc()
  )
  inference_model_2 <- beautier::create_inference_model(
    site_model = beautier::create_hky_site_model(),
    mcmc = beautier::create_nested_sampling_mcmc()
  )
  inference_models <- list(inference_model_1, inference_model_2)

  # Need the BEAST2 binary's path
  beast2_options <- beastier::create_beast2_options(
    beast2_path = beastier::get_default_beast2_bin_path()
  )

  # Need as much beast2_optionses as inference models
  beast2_optionses <- list(beast2_options, beast2_options)

  df <- est_marg_liks_from_models(
    fasta_filename,
    inference_models = inference_models,
    beast2_optionses = beast2_optionses,
    epsilon = 1e7
  )

  # Testing the results
  library(testthat)
  expect_true(is.data.frame(df))
  expect_true("site_model_name" \%in\% colnames(df))
  expect_true("clock_model_name" \%in\% colnames(df))
  expect_true("tree_prior_name" \%in\% colnames(df))
  expect_true("marg_log_lik" \%in\% colnames(df))
  expect_true("marg_log_lik_sd" \%in\% colnames(df))
  expect_true("weight" \%in\% colnames(df))

  expect_true(is.factor(df$site_model_name))
  expect_true(is.factor(df$clock_model_name))
  expect_true(is.factor(df$tree_prior_name))
  expect_true(!is.factor(df$marg_log_lik))
  expect_true(!is.factor(df$marg_log_lik_sd))
  expect_true(!is.factor(df$weight))

  expect_true(all(df$weight >= 0.0))
  expect_true(all(df$weight <= 1.0))
  expect_true(sum(df$marg_log_lik < 0.0, na.rm = TRUE) > 0)
  expect_true(sum(df$marg_log_lik_sd > 0.0, na.rm = TRUE) > 0)
}
\author{
Richel J.C. Bilderbeek
}
