% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_bayes_factor.R
\name{interpret_bayes_factor}
\alias{interpret_bayes_factor}
\title{From [1] H. Jeffreys (1961). The Theory of Probability (3rd ed.). Oxford. p. 432}
\usage{
interpret_bayes_factor(bayes_factor)
}
\description{
From [1] H. Jeffreys (1961). The Theory of Probability (3rd ed.). Oxford. p. 432
}
