% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbette_report.R
\name{mcbette_report}
\alias{mcbette_report}
\title{Create a \link{mcbette} report, to be used when reporting bugs}
\usage{
mcbette_report()
}
\description{
Create a \link{mcbette} report, to be used when reporting bugs
}
