% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mcbette_state.R
\name{set_mcbette_state}
\alias{set_mcbette_state}
\title{Set the \link{mcbette} state.}
\usage{
set_mcbette_state(
  mcbette_state,
  beast2_folder = beastier::get_default_beast2_folder()
)
}
\arguments{
\item{mcbette_state}{the \link{mcbette} state,
which is a \link{list} with the following elements:
\itemize{
  \item{
    beast2_installed
      \link{TRUE} if BEAST2 is installed,
      \link{FALSE} otherwise
  }
  \item{
     ns_installed
      \link{NA} if BEAST2 is not installed.
      \link{TRUE} if the BEAST2 NS package is installed
      \link{FALSE} if the BEAST2 NS package is not installed
  }
}}

\item{beast2_folder}{the folder where the BEAST2 is installed.
Note that this is not the folder where the BEAST2 executable is installed:
the BEAST2 executable is in a subfolder.
Use \link[beastier]{get_default_beast2_folder}
  to get the default BEAST2 folder.
Use \link[beastier]{get_default_beast2_bin_path}
  to get the full path to the default BEAST2 executable.
Use \link[beastier]{get_default_beast2_jar_path}
  to get the full path to the default BEAST2 jar file.}
}
\description{
Set the \link{mcbette} state.
}
\examples{
mcbette_state <- get_mcbette_state()
mcbette_state$beast2_installed <- TRUE
mcbette_state$ns_installed <- TRUE
\dontrun{
  set_mcbette_state(mcbette_state)
}
}
\seealso{
\itemize{
  \item Use \link{get_mcbette_state} to
    get the current \link{mcbette} state
  \item Use \link{check_mcbette_state} to
    check the current \link{mcbette} state
}
}
